+++ manifest.yml
@@ -60,6 +60,10 @@ mappings:
       handler: predictionMarketSoldCompleteSet
     - event: swaps.PoolCreate
       handler: swapPoolCreated
+    - event: swaps.SwapExactAmountIn
+      handler: swapExactAmountIn
+    - event: swaps.SwapExactAmountOut
+      handler: swapExactAmountOut
     - event: balances.Endowed
       handler: balancesEndowed
     - event: balances.DustLost
